Abstract ApiEvidenceSource + getEvidenceFromSources to use in multiple tracks
